requirements.txt
sox
python_speech_features
librosa
numpy
pandas
scipy
seaborn
matplotlib
soundfile
jiwer
wave
scipy
flask
Pillow
tensorflow-gpu==1.14.0
keras==2.3.1